build: ml: bug 26359: moving buildctl directory out of source dirs
Moving the buildctl directories out of the module source directories
at the request of the developers.

Created a flat directory structure in build/3.x/buildctl (one directory
per build module) to contain the buildctl scripts.                    

Aside from moving the files, the major change needed was to embed
the path to the source directory for each module into the buildctl
scripts.                                                          

Modified the releasbuild.sh script to call the buildctl.sh scripts
from the new location.  The script simplified a bit, since the    
path to the src directories no longer needed to be referenced in
that script.   

[git-p4: depot-paths = "//depot/software/smrtanalysis/bioinformatics/lib/python/pbcore/": change = 148947]