Add AlignmentSet Scatter by ZMW
This change adds a new alignmentset scatter task, which in contrast to
other AlignmentSet scatter gather tasks, such as
"scatter_alignments_reference.py" and
"scatter_alignments_reference_basemods.py" scatters an alignmentset by
ZMW rather than by sections of the contig they align to.

Scattering by ZMWs is an appropriate mechanisms for our HMM model
fitting, where the critical question is what are the error rates over a
ZMW, and not over a section of reads that align to a reference.

To enable this change, I added a new python file,
"scatter_alignments_reference_arrow.py", which shows features of both
code re-use and code duplication.  This file was motivated by the files
"scatter_alignments_reference.py" and
"scatter_alignments_reference_basemods.py", which in my summation appear
to be files that are only distinguished by their task names, which I
assumed was a workaround for pbsmrtpipe issue #28.  Like
"scatter_alignments_reference_basemods.py", the new file uses the parser
contract in "scatter_alignments_reference.py" so re-uses code, but
duplicates some of the methods in it so as to make minor changes to the
scattering method and also makes some other changes to enable it to work
with ZMW ranges instead of contig ranges.

This change is expected to enable enzymology while having no negative
effect on production code.  A unit test has been added to ensure it
works properly and will continue to do so going forward.